+++ contrib/expand-ids.ts
@@ -1,40 +1,16 @@
 import fs from "node:fs";
 import process from "node:process";
-import { TangutCharacterRanges as ranges } from "./constants.ts";
+import { TangutCharacterRanges, TangutComponentRanges } from "./constants.ts";
+import { parseIDSMap, formatIDSMap } from "./util.ts";
+import type { IDSMap } from "./util.ts";
 
-type IDSMap = Map<string, string>;
-
-function formatIDSMap(idsMap: IDSMap): string {
-  let output = "";
-  for (const [key, value] of idsMap) {
-    output += `${key
-      .codePointAt(0)!
-      .toString(16)
-      .toUpperCase()}\t${key}\t${value}\n`;
-  }
-  return output;
-}
-
-function parseIDSMap(input: string): IDSMap {
-  const idsMap: IDSMap = new Map();
-  const lines = input.split("\n");
-  for (const line of lines) {
-    const parts = line.split("\t");
-    if (parts.length === 3) {
-      const key = parts[1];
-      const value = parts[2];
-      idsMap.set(key, value);
-    }
-  }
-  return idsMap;
-}
-
-function expandIDS(idsMap: IDSMap, ranges: [number, number][]) {
+function expandIDS(idsMap: IDSMap, type: "character" | "component"): IDSMap {
+  const ranges = type === "character" ? TangutCharacterRanges : TangutComponentRanges;
   for (const [start, end] of ranges) {
     for (let codePoint = start; codePoint <= end; codePoint++) {
       const char = String.fromCodePoint(codePoint);
       if (!idsMap.has(char)) {
-        idsMap.set(char, ""); // Add with empty IDS if not present
+        idsMap.set(char, type === "character" ? "" : char);
       }
     }
   }
@@ -52,14 +28,15 @@ function expandIDS(idsMap: IDSMap, ranges: [number, number][]) {
 
 
 function main() {
-  if (process.argv.length < 3) {
-    console.error("Usage: node ./scripts/expand-ids.ts <ids_file>");
+  if (process.argv.length < 4) {
+    console.error("Usage: node ./scripts/expand-ids.ts <--character|--component> <ids_file>");
     process.exitCode = 1;
   }
-  const idsFile = fs.readFileSync(process.argv[2], "utf-8");
+  const type = process.argv[2].slice(2);
+  const idsFile = fs.readFileSync(process.argv[3], "utf-8");
   const idsMap = parseIDSMap(idsFile);
 
-  expandIDS(idsMap, ranges);
+  expandIDS(idsMap, type);
   const output = formatIDSMap(idsMap);
 
   fs.writeFileSync(process.argv[2], output, "utf-8");
